Initial code to make the bot work
Added the basics to enable a communication with the bot.
Updated .gitignore
Added a functionality to let the bot store the token and remember it.
Added test-command to test if the bot can answer to set keywords.
Added little information to the readme.
Added output if no set keyword got hit. (error output)